Refactor knowledge artifact transpilation to build time
