+++ src/app/knowledge/_lib/doc.ts
@@ -1,0 +1,24 @@
+export function buildHTMLDoc(html: string): string {
+  const trimmed = html.trim()
+  if (/^<!doctype|^<html/i.test(trimmed)) return html
+  return `<!DOCTYPE html><html><head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"></head><body>${html}</body></html>`
+}
+
+export function buildJSXDoc(transformedCode: string): string {
+  return `<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <script src="https://cdn.tailwindcss.com"></script>
+  <script crossorigin src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
+  <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
+</head>
+<body>
+  <div id="root"></div>
+  <script>
+${transformedCode}
+  </script>
+</body>
+</html>`
+}
